refactor(Position): Position 클래스를 player 패키지로 이동
