+++ src/test/java/domain/player/PositionTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.player;
 
+import domain.player.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
